Split pipeline into different modules
